+++ lib/glimmer/tk/menu_proxy.rb
@@ -71,8 +71,9 @@ module Glimmer
             @tk = ::TkSysMenu_Help.new(@parent_proxy.tk)
           elsif @parent_proxy.parent_proxy.is_a?(ToplevelProxy) && OS.mac? && window?
             @tk = ::Tk::TkSysMenu_Window.new(@parent_proxy.tk)
-          elsif @parent_proxy.parent_proxy.is_a?(ToplevelProxy) && OS.windows? && system?
-            @tk = ::TkSysMenu_System.new(@parent_proxy.tk)
+          # Windows system menu does not seem to work
+#           elsif @parent_proxy.parent_proxy.is_a?(ToplevelProxy) && OS.windows? && system?
+#             @tk = ::TkSysMenu_System.new(@parent_proxy.tk)
           else
             tk_widget_class = self.class.tk_widget_class_for(@keyword)
             @tk = tk_widget_class.new(@parent_proxy.tk)